services/property/requirements.txt
fastapi
gunicorn
uvicorn
psycopg2-binary
sqlalchemy
alembic
pyjwt
python-multipart
python-jose[cryptography]
pydantic[dotenv]